refactor(service): extract HealthTarget into service/health.py
Split the post-install health-poll value object out of installer.py so
the module drops from 300 lines (at the oo_score threshold, zero
headroom) to 226. HealthTarget owns host/port/token resolution for the
daemon the backend actually started; installer.py imports it and keeps
the poll loop in _verify_serving.

Behavior-preserving Split Module (PY-RF-3): the class moves, installer
wires to the new module, the old definition is deleted, and its unit
tests move to tests/test_service_health.py. Installer tests repoint the
VoxClientSync patch to punt_vox.service.health, where client() now lives.

No backwards-compat shim (PL-PP-1).
