+++ tests/test_service_installer.py
@@ -3,7 +3,6 @@
 
 from __future__ import annotations
 
-import ipaddress
 import stat as stat_mod
 from pathlib import Path
 from unittest.mock import MagicMock, patch
@@ -11,10 +10,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 
 from punt_vox.client import VoxdConnectionError, VoxdProtocolError
-from punt_vox.service.installer import (
-    ServiceInstaller,
-    _HealthTarget,  # pyright: ignore[reportPrivateUsage]
-)
+from punt_vox.service.installer import ServiceInstaller
 from punt_vox.service.launchd import LaunchdBackend
 from punt_vox.service.process import DEFAULT_PORT, ProcessManager
 from punt_vox.service.systemd import SystemdBackend
@@ -197,7 +193,7 @@ def test_install_reports_running_when_daemon_healthy(
     healthy = MagicMock()
     healthy.health.return_value = {"status": "ok"}
     monkeypatch.setattr(
-        "punt_vox.service.installer.VoxClientSync",
+        "punt_vox.service.health.VoxClientSync",
         lambda **_kwargs: healthy,
     )
 
@@ -237,7 +233,7 @@ def test_install_raises_when_daemon_never_healthy(
     down = MagicMock()
     down.health.side_effect = VoxdConnectionError("connection refused")
     monkeypatch.setattr(
-        "punt_vox.service.installer.VoxClientSync",
+        "punt_vox.service.health.VoxClientSync",
         lambda **_kwargs: down,
     )
 
@@ -457,102 +453,6 @@ def test_install_cleans_stale_user_unit_before_systemd_stop(
 
 
 # ---------------------------------------------------------------------------
-# _HealthTarget — host/port derivation for the post-install poll
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.parametrize("platform", ["macos", "linux"])
-def test_health_target_pins_installed_default_port(
-    platform: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """The poll port is DEFAULT_PORT regardless of a stray VOXD_PORT env."""
-    monkeypatch.setenv("VOXD_PORT", "9999")
-    assert _HealthTarget(platform).port == DEFAULT_PORT
-
-
-@pytest.mark.parametrize("platform", ["macos", "linux"])
-def test_health_target_unset_bind_maps_to_loopback(
-    platform: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """An unset VOXD_BIND resolves the loopback health host on both backends."""
-    monkeypatch.delenv("VOXD_BIND", raising=False)
-    assert _HealthTarget(platform).host == "127.0.0.1"
-
-
-# Construct the unspecified addresses rather than spelling "0.0.0.0" as a
-# literal, which trips ruff S104 (bind-all-interfaces) even in a test value.
-_IPV4_WILDCARD = str(ipaddress.IPv4Address(0))
-_IPV6_WILDCARD = str(ipaddress.IPv6Address(0))
-
-
-@pytest.mark.parametrize("platform", ["macos", "linux"])
-@pytest.mark.parametrize(
-    "wildcard",
-    [_IPV4_WILDCARD, _IPV6_WILDCARD, f"  {_IPV4_WILDCARD}  "],
-)
-def test_health_target_wildcard_bind_maps_to_loopback(
-    wildcard: str,
-    platform: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """Wildcard (unspecified) binds resolve loopback — voxd accepts it there."""
-    monkeypatch.setenv("VOXD_BIND", wildcard)
-    assert _HealthTarget(platform).host == "127.0.0.1"
-
-
-@pytest.mark.parametrize("platform", ["macos", "linux"])
-def test_health_target_concrete_bind_used_directly(
-    platform: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """A concrete, safe bind is polled directly on both backends."""
-    monkeypatch.setenv("VOXD_BIND", "192.168.1.50")
-    assert _HealthTarget(platform).host == "192.168.1.50"
-
-
-@pytest.mark.parametrize("platform", ["macos", "linux"])
-def test_health_target_hostname_bind_used_directly(
-    platform: str,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """A non-IP bind value (hostname) is polled as given, not loopback."""
-    monkeypatch.setenv("VOXD_BIND", "voxd.internal")
-    assert _HealthTarget(platform).host == "voxd.internal"
-
-
-def test_health_target_systemd_rejected_bind_maps_to_loopback(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """A VOXD_BIND rejected by safe_systemd_value resolves to loopback (linux).
-
-    ``SystemdBackend.unit_content`` embeds ``VOXD_BIND`` only when
-    ``safe_systemd_value`` accepts it. A trailing newline is rejected, so the
-    unit drops the variable and voxd binds its DEFAULT_HOST loopback. The
-    health host must match that -- 127.0.0.1 -- not the rejected raw address,
-    or the poll false-fails a healthy daemon.
-    """
-    monkeypatch.setenv("VOXD_BIND", "192.168.1.50\n")
-    assert not SystemdBackend.safe_systemd_value("192.168.1.50\n")
-    assert _HealthTarget("linux").host == "127.0.0.1"
-
-
-def test_health_target_launchd_does_not_apply_systemd_gate(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    """launchd embeds VOXD_BIND verbatim, so the systemd gate must not apply.
-
-    The same value the systemd gate rejects is passed through by launchd,
-    which has no safety gate. The health host must therefore resolve the
-    address (whitespace-normalized), not loopback -- the resolution diverges
-    by backend exactly as the units do.
-    """
-    monkeypatch.setenv("VOXD_BIND", "192.168.1.50\n")
-    assert _HealthTarget("macos").host == "192.168.1.50"
-
-
-# ---------------------------------------------------------------------------
 # install() — health poll targets the exact installed daemon
 # ---------------------------------------------------------------------------
 
@@ -588,7 +488,7 @@ def test_install_health_poll_pins_installed_port_over_env(
         client.health.return_value = {"status": "ok"}
         return client
 
-    monkeypatch.setattr("punt_vox.service.installer.VoxClientSync", _factory)
+    monkeypatch.setattr("punt_vox.service.health.VoxClientSync", _factory)
 
     inst = ServiceInstaller()
     inst.install()
@@ -627,7 +527,7 @@ def test_install_health_poll_retries_transient_protocol_error(
         {"status": "ok"},
     ]
     monkeypatch.setattr(
-        "punt_vox.service.installer.VoxClientSync",
+        "punt_vox.service.health.VoxClientSync",
         lambda **_kwargs: client,
     )
 
@@ -684,7 +584,7 @@ def test_install_health_poll_uses_run_file_token_not_shell_env(
         client.health.return_value = {"status": "ok"}
         return client
 
-    monkeypatch.setattr("punt_vox.service.installer.VoxClientSync", _factory)
+    monkeypatch.setattr("punt_vox.service.health.VoxClientSync", _factory)
 
     inst = ServiceInstaller()
     inst.install()
@@ -726,7 +626,7 @@ def test_install_health_poll_token_none_when_run_file_absent(
         client.health.return_value = {"status": "ok"}
         return client
 
-    monkeypatch.setattr("punt_vox.service.installer.VoxClientSync", _factory)
+    monkeypatch.setattr("punt_vox.service.health.VoxClientSync", _factory)
 
     inst = ServiceInstaller()
     inst.install()
